added motivation page & renumbered accordingly
